requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
requests==2.31.0
streamlit==1.35.0
sentence-transformers==2.6.1
faiss-cpu==1.7.4
langchain==0.2.0
PyMuPDF==1.24.2
pdfplumber==0.10.3
numpy==1.26.4
